+++ RayTracer/RayTracer/Color.h
@@ -6,14 +6,14 @@ class Color
 {
 public:
 	Color();
-	Color(uint8_t n_r, uint8_t n_g, uint8_t n_b);
+	Color(int n_r, int n_g, int n_b);
 
 	Color operator* (float s);
 	Color operator+ (Color c);
 	Color operator* (Color c);
 
-	uint8_t r;
-	uint8_t g;
-	uint8_t b;
+	int r;
+	int g;
+	int b;
 };
